refactor(test): split e2e tests into SC and QD files, remove conftest monkey-patch
Split test_e2e_quam_construction.py into separate test files for
superconducting and quantum dot architectures. Remove WiringLineType
monkey-patch from conftest.py to ensure true end-to-end interoperability
testing against py-qua-tools.
